# backend/doc/swagger-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JS Full Stack Backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/room/create": {
      "post": {
        "tags": [
          "room"
        ],
        "description": "Create a new room",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "otherUsers"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "format": "string"
                  },
                  "otherUsers": {
                    "type": "array",
                    "description": "Other users (besides the one calling the route) to give access to the route",
                    "items": {
                      "$ref": "#/components/schemas/MongoId"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the id of the room that was created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "roomId"
                  ],
                  "properties": {
                    "roomId": {
                      "$ref": "#/components/schemas/MongoId"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, details are returned, can be because of: MissingToken, BadAuthType (ex: Basic instead of Bearer)"
          },
          "401": {
            "description": "Bad token (not created by this server or expired)"
          },
          "404": {
            "description": "Any of the user are not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "missingUsers"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "missingUsers": {
                      "type": "array",
                      "description": "Users that could not be found",
                      "items": {
                        "$ref": "#/components/schemas/MongoId"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/room/info/{roomId}": {
      "get": {
        "tags": [
          "room"
        ],
        "description": "Get info on a room",
        "parameters": [
          {
            "in": "path",
            "name": "roomId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MongoId"
            }
          },
          {
            "in": "query",
            "name": "getLastPost",
            "schema": {
              "type": "boolean"
            },
            "description": "If present, the last post in the room will be returned"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns info on the room",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "users",
                    "posts"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "format": "string"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MongoId"
                      }
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MongoId"
                      }
                    },
                    "lastPost": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, details are returned, can be because of: MissingToken, BadAuthType (ex: Basic instead of Bearer)"
          },
          "401": {
            "description": "Bad token (not created by this server or expired) or user doesn't have access to the room"
          },
          "404": {
            "description": "The user or the room was not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MongoId": {
        "type": "string",
        "format": "uuid",
        "example": "632999f5bb5e77c72e9bf905"
      },
...
      "Post": {
        "type": "object",
        "required": [
          "user",
          "room",
          "content",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "user": {
            "type": "string",
            "format": "uuid",
            "example": "632999f5bb5e77c72e9bf905"
          },
          "room": {
            "type": "string",
            "format": "uuid",
            "example": "632999f5bb5e77c72e9bf905"
          },
          "content": {
            "type": "string",
            "format": "string",
            "description": "Content of the post"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2017-07-21T17:32:28Z"
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "basicAuth": [],
      "bearerAuth": []
    }
  ]
}